requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
google-generativeai==0.6.0
python-dotenv==1.0.1
pyperclip==1.8.2